Fix mobile bug, rename files
