+++ src/pages/home-page/home-page.css
@@ -1,6 +1,5 @@
 .homepage-intro-content {
     flex: 1;
-
     display: flex;
     flex-direction: column;
     justify-content: space-around;
@@ -17,6 +16,7 @@
     display: flex;
     flex-direction: column;
     flex: 1;
+
 }
 
 .homepage-instagram-section {
@@ -48,5 +48,10 @@
 
     .read-more:nth-of-type(1) {
         padding-bottom: 50px;
+    }
+
+    .read-more h1 {
+        word-wrap: break-word;
+        font-size: 24px;
     }
 }
